Add symbol references and definition information
This particular implementation is rather scuffed, but since this
is a simple language I am leaving it be. Just note that students
should not take example from this symbol implementation.

The way we do references here is we perform an extra pass over
the AST function definitions and then the parameters every time
we encounter a variable in an expression. We can then associate
each instance of a variable with its associated definition.

The normal way to do this would be to create Definition and
Reference passes in order to keep track of all that information
within a symbol table. Do that if you are attempting to write
a more complicated language.